Create README.md
As C is the basic High level programming language, which gives you the fundamental idea about what a programming language is. So you should know the C programming language. As far as career in C programming is concerned, all embedded system programs, all device drivers are written in C language. so You will get a lot of jobs in C programming.